% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sorters.R
\name{add_answer_data}
\alias{add_answer_data}
\title{Add additional answer data}
\usage{
add_answer_data(
  data,
  codebook,
  information = c("order", "option"),
  use_codebook = TRUE,
  token_name = "NETTSKJEMA_API_TOKEN",
  ...
)
}
\arguments{
\item{data}{tibble. Data retrieved from nettskjema}

\item{codebook}{codebook object retreived by \code{\link{nettskjema_get_codebook}}}

\item{information}{character vector of information to add.
One or more of "order", "option", "correct" "preselected".}

\item{use_codebook}{logical. Use codebook for retrieving answers (default TRUE). If you have not
set up the codebook in nettskjema, then turning this off will return the data as is.}

\item{token_name}{character. Name to give the token, defaults to 'NETTSKJEMA_API_TOKEN'}

\item{...}{arguments passed to \code{\link[httr]{GET}}}
}
\value{
tibble with added columns
}
\description{
The answers in the nettskjema forms
have some extra information attached
that is not returned by default.
This is for instance information on
the order of the answer options, the true
text, whether the answer is correct or
if it is preselected. This function makes
it possible to retrieve and add this
information to the nettskjema data, using
the codebook.
}
