% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_code_string.R
\name{add_code_string}
\alias{add_code_string}
\title{Add columns with the choice options rather than
their code}
\usage{
add_code_string(data, codebook, suffix = "choice")
}
\arguments{
\item{data}{nettskjema form data}

\item{codebook}{codebook from nettskjema}

\item{suffix}{suffix for columns with choice options
rather than code}
}
\description{
Add columns with the choice options rather than
their code
}
