% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_auth.R
\name{api_user_create}
\alias{api_user_create}
\title{TODO: Add possibility to supply username and description through function}
\usage{
api_user_create()
}
\description{
TODO: Add possibility to supply username and description through function
}
