% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema-codebook.R
\name{codebook}
\alias{codebook}
\title{Constructor for nettskjema codebook}
\usage{
codebook(meta_data, form_id)
}
\arguments{
\item{meta_data}{form meta-data from \code{\link{nettskjema_get_meta}}}

\item{form_id}{integer. Number of the form to retrieve}
}
\value{
long format tibble with meta-data for a specific nettskjema
}
\description{
Function works on meta-data downloaded from a form.
It will return a data.frame in long format with
meta-data information from a specific nettskjema form.
}
\examples{
\dontrun{
meta_110000 <- nettskjema_get_meta(110000)
codebook(meta_110000)
}
}
