% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_add_multiplechoice.R
\name{codebook_add_multiplechoice}
\alias{codebook_add_multiplechoice}
\title{Add indication if question allowed
multiple choice selection to codebook}
\usage{
codebook_add_multiplechoice(codebook, data)
}
\arguments{
\item{codebook}{codebook from nettskjema}

\item{data}{nettskjema form data}
}
\description{
Add indication if question allowed
multiple choice selection to codebook
}
