% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_clean.R
\name{codebook_clean}
\alias{codebook_clean}
\title{Make a data.frame of a nettskjema codebook}
\usage{
codebook_clean(codebook)
}
\arguments{
\item{codebook}{codebook}
}
\value{
a [tibble][tibble::tibble()] type data.frame
}
\description{
\code{codebook_clean} wrangles the codebook into a tall data.frame
}
\author{
Athanasia Mowinckel
}
