% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{element_details}
\alias{element_details}
\title{Grabs information about nettskjema elements}
\usage{
element_details(type, el)
}
\arguments{
\item{type}{type of element}

\item{el}{nettskjema element}
}
\description{
Grabs information about nettskjema elements
}
