% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{element_exists}
\alias{element_exists}
\title{Title}
\usage{
element_exists(el, fields)
}
\arguments{
\item{el}{nettskjema element}

\item{fields}{character vector of nettskjema fields}
}
\description{
Title
}
