% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-codebook.R
\name{has_codebook}
\alias{has_codebook}
\title{Check if form has codebook}
\usage{
has_codebook(form_id)
}
\arguments{
\item{form_id}{integer. Number of the form to retrieve}
}
\value{
logical is codebook is turned on
}
\description{
Codebook is by default turned off
in Nettskjema, but best practices in handling
data is to have it on. This function
checks if the codebook of a form has been activated
or not
}
\examples{
\dontrun{
has_codebook(110000)
}
}
