% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_auth.R
\name{nettskjema_api}
\alias{nettskjema_api}
\title{nettskjema api connection}
\usage{
nettskjema_api(path, token_name, ...)
}
\arguments{
\item{path}{path connection}

\item{token_name}{token name for lookup}

\item{...}{arguments passed to httr::GET}
}
\value{
an httr reponse
}
\description{
nettskjema api connection
}
