% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_auth.R
\name{nettskjema_find_ip}
\alias{nettskjema_find_ip}
\title{Find your current IP}
\usage{
nettskjema_find_ip(version = c("v4", "v6"))
}
\arguments{
\item{version}{string. ip version to check.
one of either "v4" or "v6".}
}
\description{
The Nettskjema v2 API requires a token
to be registered with IP addresses.
This function returns the current IP address.
}
\details{
If you are working through a VPN, the IP address
will not be returned as the Nettskjema API sees it.
We are looking for solutions to this.
}
