% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_get.R
\name{nettskjema_get_meta}
\alias{nettskjema_get_meta}
\title{Get metadata for a form}
\usage{
nettskjema_get_meta(form_id, token_name = "NETTSKJEMA_API_TOKEN", ...)
}
\arguments{
\item{form_id}{integer. Number of the form to retrieve}

\item{token_name}{character. Name to give the token, defaults to 'NETTSKJEMA_API_TOKEN'}

\item{...}{arguments passed to \code{\link[httr]{GET}}}
}
\value{
list of class nettskjema_meta_data
}
\description{
With the given API token, will retrieve
the meta-data connected to a form.
}
\examples{
\dontrun{
meta_110000 <- nettskjema_get_meta(110000)

}
}
