% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema-labels.R
\name{nettskjema_label_vars}
\alias{nettskjema_label_vars}
\title{Add variable labels}
\usage{
nettskjema_label_vars(data, checkbox_type = "string", metadata = NULL)
}
\arguments{
\item{data}{tibble. Data retrieved from nettskjema}

\item{checkbox_type}{string of either "string" (default), "list" or "columns" for how to handle checkbox answers}

\item{metadata}{Optional, raw metadata as returned from Nettskjema API.}
}
\value{
tibble with label attributed to columns
}
\description{
The Nettskjema forms come with meta-data that
allows for nicer handling of label information.
This function will use the metadata and and apply
labels to the variables.
}
