% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema-attachments.R
\name{nettskjema_list_attachments}
\alias{nettskjema_list_attachments}
\title{List Nettskjema attachments}
\usage{
nettskjema_list_attachments(
  submission_id,
  token_name = "NETTSKJEMA_API_TOKEN",
  ...
)
}
\arguments{
\item{submission_id}{integer. Unique identifier of a specific Nettskjema submission.}

\item{token_name}{character. Name to give the token, defaults to 'NETTSKJEMA_API_TOKEN'}

\item{...}{arguments passed to \code{\link[httr]{GET}}}
}
\value{
a tibble with information on attachments available.
}
\description{
Each unique submission ID in Nettskjema
may have some attachments. These can be listed
using this function, which will provide the
information on where in the Nettskjema API the
files are located, their original file names,
and a suggested standardised file names for
tidier data output.
}
\examples{
\dontrun{
submission_id <- c(22222, 1232, 21555)
nettskjema_list_attachments(submission_id)

}
}
