% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_auth.R
\name{nettskjema_renviron_edit}
\alias{nettskjema_renviron_edit}
\title{Open .Renviron file for editing}
\usage{
nettskjema_renviron_edit()
}
\description{
Will open the .Renviron file for editing.
In this file you set environment variables
that might be required for using certain
command line arguments through functions.
In this package the .Renviron stores the
API access tokens necessary to communicate
securely with nettskjema.
}
