% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema-attachments.R
\name{nettskjema_save_attachment}
\alias{nettskjema_save_attachment}
\title{Save Nettskjema attachments to file}
\usage{
nettskjema_save_attachment(path, output, token_name, ...)
}
\arguments{
\item{path}{Nettskjema API path where the attachment is}

\item{output}{output file name}

\item{token_name}{character. Name to give the token, defaults to 'NETTSKJEMA_API_TOKEN'}

\item{...}{arguments passed to \code{\link[httr]{GET}}}
}
\value{
nothing. saves attachments to file.
}
\description{
The Nettskjema forms has an option
to upload attachments with forms.
These can be retrieved by this function.
Recommended workflow is to first call
\code{\link{nettskjema_list_attachments}},
and use the output of this to pass along
to the function. This function is called
by \code{\link{nettskjema_get_form_attachments}},
but you can use it to define your own output
file names.
}
\examples{
\dontrun{
submission_id <- c(22222, 1232, 21555)
attach_dt <- nettskjema_list_attachments(submission_id)
nettskjema_save_attachment(path = attach_dt$path, output = attach_dt$standardised)
}
}
