% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_auth.R
\name{nettskjema_token2renviron}
\alias{nettskjema_token2renviron}
\title{Write token to .Renviron}
\usage{
nettskjema_token2renviron(
  token,
  token_name = "NETTSKJEMA_API_TOKEN",
  action = c("create", "overwrite", "delete")
)
}
\arguments{
\item{token}{character. Token generated in the UiO portal \code{api_user_create}}

\item{token_name}{character. Name to give the token, defaults to 'NETTSKJEMA_API_TOKEN'}

\item{action}{character. One of three actions: 'create', 'overwrite' or 'delete'.
Defaults to 'create'.}
}
\description{
This function will write a token
to the .Renviron file with the
name you provide.
}
