% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_auth.R
\name{nettskjema_token_expiry}
\alias{nettskjema_token_expiry}
\title{Check nettskjema API token expiry date}
\usage{
nettskjema_token_expiry(token_name = "NETTSKJEMA_API_TOKEN")
}
\arguments{
\item{token_name}{Given name of the token}
}
\value{
character date string
}
\description{
Given an API token, will retrieve the expiry
date of the said token.
}
