% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjemar-package.R
\docType{package}
\name{nettskjemar-package}
\alias{nettskjemar}
\alias{nettskjemar-package}
\title{nettskjemar: Connect to the 'nettskjema.no' API of the University of Oslo}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Enables users to retrieve data, meta-data, and codebooks from \url{https://nettskjema.no/}. The data from the API is richer than from the online data portal. This package is not developed by the University of Oslo IT. Mowinckel (2021) \doi{10.5281/zenodo.4745481}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CAPRO-UiO/nettskjemar}
  \item \url{https://zenodo.org/badge/latestdoi/206264675}
  \item \url{https://github.com/capro-uio/nettskjemar}
  \item Report bugs at \url{https://github.com/capro-uio/nettskjemar/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Other contributors:
\itemize{
  \item Trym Nohr Fjørtoft \email{t.n.fjortoft@arena.uio.no} (\href{https://orcid.org/0000-0002-7654-4587}{ORCID}) [contributor]
}

}
\keyword{internal}
