% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-browser.R
\name{ns_api_docs}
\alias{ns_api_docs}
\title{Access Nettskjema API Documentation}
\usage{
ns_api_docs()
}
\value{
None. Opens a browser window.
}
\description{
Opens the browser to display the API documentation
for the Nettskjema PI.
}
\examples{
\dontrun{
ns_api_docs()
}
}
