% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-codebook.R
\name{ns_get_codebook}
\alias{ns_get_codebook}
\title{Get codebook for a form}
\usage{
ns_get_codebook(form_id, asis = FALSE)
}
\arguments{
\item{form_id}{integer. Number of the form to retrieve}

\item{asis}{logical. Whether to return the raw content from nettskjema. Default is FALSE.}
}
\value{
list of class ns_meta_data
}
\description{
With the given API token, will retrieve
a list of all the forms you have access to
}
\examples{
\dontrun{
codebook_110000 <- ns_get_codebook(110000)
}
}
