% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-attachment.R
\name{ns_get_form_attachments}
\alias{ns_get_form_attachments}
\title{Retrieve all attachments associated with a Nettskjema}
\usage{
ns_get_form_attachments(
  form_id,
  filenames = c("standardized", "original"),
  output_dir = "."
)
}
\arguments{
\item{form_id}{integer. Number of the form to retrieve}

\item{filenames}{Character of either 'standardized'
(default) or 'original' indicating which file names to use.}

\item{output_dir}{directory to output the files to}
}
\value{
invisible named logical if the file was saved successfully
}
\description{
Some Nettskjema have fields for attachments. These
can be downloaded through this function. There
are two naming convention that may be used for the
saved files' names. See output of
\code{\link{ns_list_submission_attachments}} for idea
of the names used.
}
\details{
filenames types
\describe{
 \item{"original"}{ - uses file names are they were
    uploaded to Nettskjema}
 \item{"standardized"}{ - creates file names based on
    the submission id and a counter number to uniquely
    create file names for each submission (in case there
    are more than one attachment)}
 }
}
\examples{
\dontrun{
form_id <- 1100000
ns_get_form_attachments(form_id)

# save files to specific folder
ns_get_form_attachments(form_id, output_dir = "~/Desktop")

# save using original file names
ns_get_form_attachments(form_id, filenames = "original")

}
}
