% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-attachment.R
\name{ns_list_submission_attachments}
\alias{ns_list_submission_attachments}
\title{List Nettskjema attachments for a submission}
\usage{
ns_list_submission_attachments(submission_id)
}
\arguments{
\item{submission_id}{integer. Unique identifier of a specific Nettskjema submission.}
}
\value{
a data.frame with information on attachments available.
}
\description{
Each unique submission ID in Nettskjema
may have some attachments. These can be listed
using this function, which will provide the
information on where in the Nettskjema API the
files are located, their original file names,
and a suggested standardized file names for
tidier data output.
}
\examples{
\dontrun{
submission_id <- 22222
ns_list_submission_attachments(submission_id)

}
}
