% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-auth.R
\name{ns_req_auth}
\alias{ns_req_auth}
\title{Authenticate Nettskjema request}
\usage{
ns_req_auth(
  req,
  client_id = Sys.getenv("NETTSKJEMA_CLIENT_ID"),
  client_secret = Sys.getenv("NETTSKJEMA_CLIENT_SECRET"),
  client_name = "nettskjemar"
)
}
\arguments{
\item{client_id}{Character. Retrieved from the Client portal.}

\item{client_secret}{Character. Retrieved from the Client portal.}

\item{client_name}{Character. Used to identify who has been running the commands.}
}
\description{
After creating a client in Nettskjema,
this function will retrieve the access
token needed for the remaining processes
in the package. Automatically caches the
token for more efficient API usage.
}
