% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-browser.R
\name{nskj_api_docs}
\alias{nskj_api_docs}
\title{Access Nettskjema API Documentation}
\usage{
nskj_api_docs()
}
\value{
None. Opens a browser window.
}
\description{
Opens the browser to display the API documentation
for the Nettskjema PI.
}
\examples{
\dontrun{
nskj_api_docs()
}
}
