% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-browser.R
\name{nskj_create_client}
\alias{nskj_create_client}
\title{Create a Nettskjema Client}
\usage{
nskj_create_client()
}
\value{
NULL. Opens the client creation page in a browser.
}
\description{
Opens the Nettskjema client creation URL in the default web browser.
}
\details{
Use this function to create a new client for accessing Nettskjema APIs.
}
\examples{
\dontrun{
nskj_create_client()
}
}
