% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-attachment.R
\name{nskj_get_attachment}
\alias{nskj_get_attachment}
\title{Save Nettskjema attachments to file}
\usage{
nskj_get_attachment(attachment_id, path)
}
\arguments{
\item{attachment_id}{integer. Unique ID of attachment, as found in the submission data.}

\item{path}{Nettskjema API path where the attachment is}
}
\value{
nothing. saves attachments to file.
}
\description{
The Nettskjema forms has an option
to upload attachments with forms.
These can be retrieved by this function.
Recommended workflow is to first call
\code{\link{nskj_list_attachments}},
and use the output of this to pass along
to the function. This function is called
by \code{\link{nskj_get_form_attachments}},
but you can use it to define your own output
file names.
}
\examples{
\dontrun{
#submission_id <- c(22222, 1232, 21555)
#attach_dt <- nskj_list_attachments(submission_id)
#nskj_save_attachment(path = attach_dt$path, output = attach_dt$standardized)
}
}
