% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-submission.R
\name{nskj_get_data}
\alias{nskj_get_data}
\title{Get data from a form}
\usage{
nskj_get_data(form_id, asis = FALSE, labelled = !asis)
}
\arguments{
\item{form_id}{integer. Number of the form to retrieve}

\item{asis}{logical. Whether to return the raw content from nettskjema. Default is FALSE.}

\item{labelled}{Logical. Should labelled data be
returned. Labelled data applies contextual
information from the codebook into special
attributes in the data.frame. See the
\code{labelled} package for more information.}
}
\value{
tibble data.frame
}
\description{
This function connects to a specific form and fetches all answers.
The responses are placed in a data.frame for easy use.
}
\examples{
\dontrun{

# Retrieve all data
data_110000 <- nskj_get_data(110000)

}
}
