% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-form.R
\name{nskj_get_forms}
\alias{nskj_get_forms}
\title{Get all forms you have access to}
\usage{
nskj_get_forms(asis = FALSE)
}
\arguments{
\item{asis}{logical. Whether to return the raw content from nettskjema. Default is FALSE.}
}
\value{
data.frame, if \code{asis = TRUE} returns a list
}
\description{
All users have access to specific
forms, and this function uses the
token provided to check what the
current user has access to.
}
