% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-me.R
\name{nskj_get_me}
\alias{nskj_get_me}
\title{Information about the current user}
\usage{
nskj_get_me()
}
\value{
list of information
}
\description{
Based on login credential, returns
information on the user accessing the
Nettskjema PI
}
\examples{
\dontrun{
nskj_get_me()
}
}
