% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-meta.R
\name{nskj_get_meta}
\alias{nskj_get_meta}
\title{Get metadata for a form}
\usage{
nskj_get_meta(form_id)
}
\arguments{
\item{form_id}{integer. Number of the form to retrieve}
}
\value{
list of class nskj_meta_data
}
\description{
With the given API token, will retrieve
the meta-data connected to a form.
}
\examples{
\dontrun{
meta_110000 <- nskj_get_meta(110000)

}
}
