% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-submission.R
\name{nskj_get_submission}
\alias{nskj_get_submission}
\title{Get an individual submission answer}
\usage{
nskj_get_submission(submission_id)
}
\arguments{
\item{submission_id}{integer. Unique identifier of a specific Nettskjema submission.}
}
\value{
list of all answers
}
\description{
Get an individual submission answer
}
\examples{
# example code
\dontrun{
nskj_get_submission(1100)
}
}
