% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-submission.R
\name{nskj_get_submission_pdf}
\alias{nskj_get_submission_pdf}
\title{Get an individual submission answer as a pdf}
\usage{
nskj_get_submission_pdf(
  submission_id,
  path = sprintf("\%s.pdf", submission_id)
)
}
\arguments{
\item{submission_id}{integer. Unique identifier of a specific Nettskjema submission.}

\item{path}{Character. File path.}
}
\value{
nothing, writes file to path
}
\description{
Get an individual submission answer as a pdf
}
\examples{
# example code
\dontrun{
nskj_get_submission_pdf(1100)
nskj_get_submission_pdf(1100, "~/Desktop/1100.pdf")
}
}
