% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-attachment.R
\name{nskj_list_form_attachments}
\alias{nskj_list_form_attachments}
\title{Retrieve all attachments associated with a Nettskjema}
\usage{
nskj_list_form_attachments(form_id)
}
\arguments{
\item{form_id}{integer. Number of the form to retrieve}
}
\value{
invisible named logical if the file was saved successfully
}
\description{
Some Nettskjema have fields for attachments. These
can be downloaded through this function. There
are two naming convention that may be used for the
saved files' names. See output of
\code{\link{nskj_list_submission_attachments}} for idea
of the names used.
}
\details{
filenames types
\describe{
 \item{"original"}{ - uses file names are they were
    uploaded to Nettskjema}
 \item{"standardized"}{ - creates file names based on
    the submission id and a counter number to uniquely
    create file names for each submission (in case there
    are more than one attachment)}
 }
}
\examples{
\dontrun{
form_id <- 1100000
nskj_list_form_attachments(form_id)

}
}
