% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nskj-auth.R
\name{nskj_req}
\alias{nskj_req}
\title{Basic Nettskjema httr2 request}
\usage{
nskj_req(...)
}
\arguments{
\item{...}{arguments passed to \code{nskj_auth_token}}
}
\value{
\code{httr2} request
}
\description{
Sets the API url and access token.
This is used for almost all the
functionality in this package. If
you want to use endpoints not yet
covered by functions in the package,
this is the basis function you can
build from.
}
\examples{
\dontrun{
nskj_req() |>
 httr2::req_url_path_append("form", "me") |>
 httr2::req_perform() |>
 httr2::resp_body_json()
}
}
