% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_html}
\alias{strip_html}
\title{Strip HTML from string}
\usage{
strip_html(string)
}
\arguments{
\item{string}{String to strip HTML from.}
}
\value{
character vector without html tags.
}
\description{
Strip HTML from string
}
\examples{
strip_html("<p>This string has the paragraph html tag</p>")
}
