% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{add_answer_data}
\alias{add_answer_data}
\title{Add additional answer data}
\usage{
add_answer_data(data, codebook, information = c("order", "option"))
}
\arguments{
\item{data}{nettskjema data}

\item{codebook}{codebook from same nettskjema}

\item{information}{character vector of information to add.
One or more of "order", "option", "correct" "preselected".}
}
\value{
tibble with added columns
}
\description{
The answers in the nettskjema forms
have some extra information attached
that is not returned by default.
This is for instance information on
the order of the answer options, the true
text, whether the answer is correct or
if it is preselected. This function makes
it possible to retrieve and add this
information to the nettskjema data, using
the codebook.
}
