% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-codebook.R
\name{codebook}
\alias{codebook}
\title{Constructor for nettskjema codebook}
\usage{
codebook(meta)
}
\arguments{
\item{meta}{form meta-data from \code{\link{ns_get_meta}}}
}
\value{
long format tibble with meta-data for a specific nettskjema
}
\description{
Function works on meta-data downloaded from a form.
It will return a data.frame in long format with
meta-data information from a specific nettskjema form.
}
\examples{
\dontrun{
meta_110000 <- ns_get_meta(110000)
codebook(meta_110000)
}
}
