% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Grab codebook for a form}
\usage{
codebook(meta_data)
}
\arguments{
\item{meta_data}{form meta-data from \code{\link{nettskjema_get_meta}}}
}
\value{
long format tibble
}
\description{
Function works on meta-data downloaded from a form.
It will return a data.frame in long format
}
\examples{
\dontrun{
meta_110000 <- nettskjema_get_meta(110000)
codebook(meta_110000)
}
}
