% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_get.R
\name{nettskjema_get_data}
\alias{nettskjema_get_data}
\title{Get data from a form}
\usage{
nettskjema_get_data(form_id, use_codebook = TRUE,
  token_name = "NETTSKJEMA_API_TOKEN")
}
\arguments{
\item{form_id}{integer. Number of the form to retrieve}

\item{use_codebook}{logical. Use codebook for answers, or
retrieve actualy text options.
data from}

\item{token_name}{character. Name to give the token, defaults to 'NETTSKJEMA_API_TOKEN'}
}
\value{
tibble data.frame
}
\description{
Get data from a form
}
