% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_get.R
\name{nettskjema_get_forms}
\alias{nettskjema_get_forms}
\title{Get all forms you have access to}
\usage{
nettskjema_get_forms(token_name = "NETTSKJEMA_API_TOKEN")
}
\arguments{
\item{token_name}{character. Name to give the token, defaults to 'NETTSKJEMA_API_TOKEN'}
}
\value{
list
}
\description{
With the given API token, will retrieve
a list of all the forms you have access to
TODO: make this work
}
