% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_get.R
\name{nettskjema_get_meta}
\alias{nettskjema_get_meta}
\title{Get metadata for a form}
\usage{
nettskjema_get_meta(form_id, information = c("title", "language",
  "created", "modified", "opened", "respondents", "contact", "codebook",
  "personal_data", "sensitive_data", "editors", "elements"),
  token_name = "NETTSKJEMA_API_TOKEN")
}
\arguments{
\item{form_id}{integer. Number of the form to retrieve}

\item{information}{which meta data elements to extract}

\item{token_name}{character. Name to give the token, defaults to 'NETTSKJEMA_API_TOKEN'}
}
\value{
list
}
\description{
With the given API token, will retrieve
a list of all the forms you have access to
}
