% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettskjema_auth.R
\name{nettskjema_token2renviron}
\alias{nettskjema_token2renviron}
\title{Write token to .Renviron}
\usage{
nettskjema_token2renviron(token, token_name = "NETTSKJEMA_API_TOKEN",
  overwrite = FALSE)
}
\arguments{
\item{token}{character. Token generated in the UiO portal \code{api_user_create}}

\item{token_name}{character. Name to give the token, defaults to 'NETTSKJEMA_API_TOKEN'}

\item{overwrite}{logical. Overwrite token if token by that name already exists.
Defaults to FALSE}
}
\description{
This function will write a token
to the .Renviron file with the
name you provide.
}
