% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-me.R
\name{ns_get_me}
\alias{ns_get_me}
\title{Information about the current user}
\usage{
ns_get_me()
}
\value{
list of information
}
\description{
Based on login credential, returns
information on the user accessing the
Nettskjema PI
}
\examples{
\dontshow{
vcr::insert_example_cassette("ns_get_me", package = "nettskjemar")
}
ns_get_me()
\dontshow{
vcr::eject_cassette()
}
}
