% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-meta.R
\name{ns_get_meta}
\alias{ns_get_meta}
\title{Get metadata for a form}
\usage{
ns_get_meta(form_id)
}
\arguments{
\item{form_id}{integer. Number of the form to retrieve}
}
\value{
list of class ns_meta_data
}
\description{
With the given API token, will retrieve
the meta-data connected to a form.
}
\examples{
\dontshow{
vcr::insert_example_cassette("ns_get_meta", package = "nettskjemar")
nettskjemar:::mock_auth()
}
ns_get_meta(123823)
\dontshow{
vcr::eject_cassette()
}
}
