% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns-submission.R
\name{ns_get_submission}
\alias{ns_get_submission}
\title{Get an individual submission answer}
\usage{
ns_get_submission(submission_id)
}
\arguments{
\item{submission_id}{integer. Unique identifier of a specific Nettskjema submission.}
}
\value{
list of all answers
}
\description{
Get an individual submission answer
}
\examples{
\dontshow{
vcr::insert_example_cassette("ns_get_submission", package = "nettskjemar")
nettskjemar:::mock_auth()
}
ns_get_submission(27685292)
\dontshow{
vcr::eject_cassette()
}
}
